VibeVoice/pipecat_pipeline/pyproject.toml
[project]
name = "vibevoice-pipecat-agent"
version = "0.1.0"
description = "Pipecat voice agent with Gemini STT/LLM and VibeVoice TTS"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pipecat-ai[daily,google,silero,webrtc,runner]",
    "websockets",
    "numpy",
    "aiohttp",
    "loguru",
    "python-dotenv",
    "fastapi",
    "uvicorn"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# This is a script-based project, not a package
# Just include the Python files
only-include = [
    "bot.py",
    "custom_vibevoice_tts_service.py",
    "test_local.py",
]

